fix(stepdefinitionpath): do not rely on cypress configuration for determining the stepDefinitionPath
Relying on any parts of the cypress configuration has been problematic since the beginning, it was
time to use only the configuration that we own. This is a breaking change but it should be
relatively easy to adjust to it. And for people using defaults - nothing changes. Should improve
performance especially on Windows.

BREAKING CHANGE: You need to configure the stepDefinitions if you use the nonGlobalStepDefinitions
option.

re #295 , fix #283